Fix git test
