+++ src/test/java/com/cloudogu/scmmanager/info/SourceUtilTestHelper.java
@@ -1,0 +1,31 @@
+package com.cloudogu.scmmanager.info;
+
+import com.cloudogu.scmmanager.scm.ScmManagerSource;
+import hudson.model.Run;
+import junit.framework.TestCase;
+
+import java.util.Arrays;
+import java.util.List;
+import java.util.stream.Collectors;
+
+import static org.mockito.Mockito.doReturn;
+import static org.mockito.Mockito.mock;
+
+class SourceUtilTestHelper extends TestCase {
+  static void mockSource(Run<TestJob, TestRun> run, String... urls) {
+    TestJob job = mock(TestJob.class);
+    TestSCMSourceOwner sourceOwner = mock(TestSCMSourceOwner.class);
+
+    doReturn(job).when(run).getParent();
+    doReturn(sourceOwner).when(job).getParent();
+
+    List<ScmManagerSource> sources = Arrays.stream(urls).map(SourceUtilTestHelper::createSource).collect(Collectors.toList());
+    doReturn(sources).when(sourceOwner).getSCMSources();
+  }
+
+  private static ScmManagerSource createSource(String url) {
+    ScmManagerSource scmSource = mock(ScmManagerSource.class);
+    doReturn(url).when(scmSource).getRemoteUrl();
+    return scmSource;
+  }
+}
