+++ src/test/java/com/cloudogu/scmmanager/info/TestRun.java
@@ -1,0 +1,12 @@
+package com.cloudogu.scmmanager.info;
+
+import hudson.model.Run;
+
+import javax.annotation.Nonnull;
+import java.io.IOException;
+
+public abstract class TestRun extends Run<TestJob, TestRun> {
+  protected TestRun(@Nonnull TestJob job) throws IOException {
+    super(job);
+  }
+}
