+++ src/test/java/com/cloudogu/scmmanager/info/TestSCMSourceOwner.java
@@ -1,0 +1,15 @@
+package com.cloudogu.scmmanager.info;
+
+import jenkins.model.Jenkins;
+import jenkins.scm.api.SCMSourceOwner;
+import org.jvnet.hudson.reactor.ReactorException;
+
+import javax.servlet.ServletContext;
+import java.io.File;
+import java.io.IOException;
+
+public abstract class TestSCMSourceOwner extends Jenkins implements SCMSourceOwner {
+  protected TestSCMSourceOwner(File root, ServletContext context) throws IOException, InterruptedException, ReactorException {
+    super(root, context);
+  }
+}
